+++ lib/entify.dart
@@ -10,5 +10,5 @@ export "src/datastore/errors.dart";
 export "src/datastore/key.dart";
 export "src/datastore/mutations.dart";
 export "src/datastore/query.dart";
-export "src/datastore/shell.dart";
+export "src/datastore/shell.dart" hide GenericQueryResult;
 export "src/datastore/values.dart";