+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1+3
+
+* Added `createKey` method to `EntityBridge`
+* Created `QueryResult` interface
+
 ## 0.1.1
 
 * The `bool` parameter of `toValue` became a named parameter as it is recommended.